Crate gauss_lib

This library implements Gaussian processes (GP) for use in Bayesian Regression with an aim of use in minimisation of chemical structures.

GP regression treats the unknown function values at all training points as a Gaussian Process fully characterised by some arbitrary covariance function or kernel between points $x_{1}$ and $x_{2}$.

The output is modelled as a stochastic process:

$ Y(x) = \mu(x) + Z(x)$

where $\mu$ is the mean and $Z(x)$ realization of the stochastic process

Implementation

Based of Faer to provide linear algebra subroutine

GP mean currently must be 0

GP correlation function can be arbitrary, with parameterisation

Dimensionality issue : scales as $O(n^{3})$ in size of training data

‘Focused’ regression, utilising only training points most similar to test point to reduce complexity

Reference

Gaussian Processes for Machine Learning, C. E. Rasmussen & C. K. I. Williams, 2006

Jones, M. R., et al. “Constraining Gaussian processes for grey-box acoustic emission source localisation.” Proceedings of the 29th international conference on noise and vibration engineering (ISMA 2020). 2020.

Jan N. Fuhg, Michele Marino, Nikolaos Bouklas, Local approximate Gaussian process regression for data-driven constitutive models: development and comparison with neural networks, Computer Methods in Applied Mechanics and Engineering, Volume 388, 2022, 114217, ISSN 0045-7825, https://doi.org/10.1016/j.cma.2021.114217.

arXiv:1402.0645 [cs.LG] https://doi.org/10.48550/arXiv.1402.0645

Lafage, R., (2022). egobox, a Rust toolbox for efficient global optimization. Journal of Open Source Software, 7(78), 4737, https://doi.org/10.21105/joss.04737

arXiv:2306.06475 [cond-mat.mtrl-sci] https://doi.org/10.48550/arXiv.2306.06475

Structs

GaussProcs

Creates a Gaussian Process Solver given a vector of inputs, $x$ and outputs $y$, a value for the variance $\sigma$ and an initial value for the hyperparameters $\theta$

Enums

ProcessError

Error in the GP

Traits

Kernel

Trait bounds needed for a type $\text{T}$ to be a valid input for the Gaussian processes